chore: 아휴..

+++ server/src/main/java/com/codestates/product/dto/ProductResponseDto.java
@@ -47,7 +47,7 @@ public class ProductResponseDto {
         private String description;
         private Product.ProductStatus productStatus;
         private boolean favoriteStatus;
-        private Long favoriteCount;
+        private int favoriteCount;
         private LocalDateTime creationDate;
         private LocalDateTime lastEditDate;
         private PcategoryResonseDto pcategory;
